refactor(kitchen pos) : MarginValidator 가 MenuRepository에 강하게 의존하고 있어 패키지 이동
